S10P22A407-7 feat: Add list related api for mypage

+++ backend/api/src/main/java/store/dalkak/api/user/dto/request/UserCreateSurveyResultReqDto.java
@@ -9,6 +9,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 @Builder
 @Getter
@@ -19,7 +20,7 @@ public class UserCreateSurveyResultReqDto {
     List<Long> surveyCocktails;
     Long occasionId;
     Long baseId;
-    Long alcoholContent;
-    Long sweetness;
+    Integer alcoholContent;
+    Integer sweetness;
     List<Long> surveyIngredients;
 }
